Add files via upload
I developed this Java program in order to help make business calculations in a much less frictionless way. It allows me to be able to make calculations on any blockchain crypto project with changing variables such as network difficulty, price changes, cost of power, and whether or not your selling monthly to cover overhead costs.